refactor(dashboard): extract vertical navigation component and subcomponents into dedicated directory in components folder
