+++ src/components/dashboard/layout/index.tsx
@@ -1,0 +1,1 @@
+export { VerticalNavigation } from "./vertical-navigation";
